Move experiment functionality
