+++ docs/snippets/estimate_pi/montecarlo.py
@@ -22,7 +22,7 @@ class EstimatePi(Component):
         )
 
     def summary(self):
-        if self.is_finished():
+        if self.execution.is_finished():
             print(
                 f"After {self.config.samples} samples, "
                 f"PI is approximately {self.load_data('result.json')['pi']}."
