+++ src/machinable/interface.py
@@ -1,9 +1,31 @@
 from typing import Any, List, Optional, Union
 
+import shlex
+import sys
+
+from flatten_dict import flatten
+
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    from typing_extensions import Self
+
 import os
 
-from machinable.collection import Collection
-from machinable.element import Element, resolve_custom_predicate
+from machinable import schema
+from machinable.collection import (
+    Collection,
+    ComponentCollection,
+    ElementCollection,
+)
+from machinable.element import (
+    Element,
+    belongs_to,
+    get_dump,
+    get_lineage,
+    has_many,
+    resolve_custom_predicate,
+)
 from machinable.settings import get_settings
 from machinable.types import VersionType
 from omegaconf import OmegaConf
@@ -12,6 +34,78 @@ from omegaconf import OmegaConf
 class Interface(Element):
     kind = "Interface"
     default = get_settings().default_interface
+
+    def __init__(
+        self,
+        version: VersionType = None,
+        uses: Union[None, Element, List[Element]] = None,
+        derived_from: Optional["Component"] = None,
+    ):
+        super().__init__(version=version)
+        self.__model__ = schema.Interface(
+            module=self.__model__.module,
+            config=self.__model__.config,
+            version=self.__model__.version,
+            lineage=get_lineage(self),
+        )
+        self.__model__._dump = get_dump(self)
+        self.__related__["uses"] = ElementCollection()
+        if uses:
+            self.use(uses)
+        if derived_from is not None:
+            self.__model__.derived_from = derived_from.uuid
+            self.__related__["ancestor"] = derived_from
+
+    def commit(self) -> "Component":
+        self.on_before_commit()
+
+        Storage.get().commit(self)
+
+        return self
+
+    @has_many
+    def derived() -> "ComponentCollection":
+        """Returns a collection of derived components"""
+        return Component, ComponentCollection, False
+
+    @belongs_to
+    def ancestor() -> Optional["Component"]:
+        """Returns parent component or None if component is independent"""
+        return Component
+
+    @has_many
+    def uses() -> ElementCollection:
+        return Element, ElementCollection
+
+    def to_cli(self) -> str:
+        cli = [self.module]
+        for v in self.__model__.version:
+            if isinstance(v, str):
+                cli.append(v)
+            else:
+                cli.extend(
+                    [
+                        f"{key}={shlex.quote(str(val))}"
+                        for key, val in flatten(v, reducer="dot").items()
+                    ]
+                )
+
+        return " ".join(cli)
+
+    def use(self, use: Union[Element, List[Element]]) -> "Self":
+        self._assert_editable()
+
+        if isinstance(use, (list, tuple)):
+            for _use in use:
+                self.use(_use)
+            return self
+
+        if not isinstance(use, Element):
+            raise ValueError(f"Expected element, but found: {type(use)} {use}")
+
+        self.__related__["uses"].append(use)
+
+        return self
 
     @classmethod
     def singleton(
@@ -148,6 +242,9 @@ class Interface(Element):
             return None
 
         return self.__model__._storage_instance.path(self, *append)
+
+    def on_compute_path(self) -> None:
+        pass
 
     def local_directory(
         self, *append: str, create: bool = False
